+++ pages/api/nfts/index.ts
@@ -1,8 +1,8 @@
 import { withSentry } from '@sentry/nextjs'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
-import dbConnect from '../../../lib/connectMongo'
 import { update } from '../../../lib/http'
+import dbConnect from '../../../lib/mongoClient'
 import redisClient from '../../../lib/redisClient'
 import { INft, INftDoc, Nft } from '../../../models/nft.model'
