+++ pages/api/projects/index.ts
@@ -1,9 +1,9 @@
 import { withSentry } from '@sentry/nextjs'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
-import dbConnect from '../../../lib/connectMongo'
 import { update } from '../../../lib/http'
 import logger from '../../../lib/logger'
+import dbConnect from '../../../lib/mongoClient'
 import redisClient, { connectRedis, DEFAULT_EXPIRY, disconnectRedis } from '../../../lib/redisClient'
 import { IProject, IProjectDoc, Project } from '../../../models/project.model'
