+++ pages/api/stems/index.ts
@@ -1,7 +1,7 @@
 import { withSentry } from '@sentry/nextjs'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
-import dbConnect from '../../../lib/connectMongo'
+import dbConnect from '../../../lib/mongoClient'
 import redisClient from '../../../lib/redisClient'
 import type { IStemDoc } from '../../../models/stem.model'
 import { Stem } from '../../../models/stem.model'
