+++ pages/api/users/collaborator/[id].ts
@@ -1,7 +1,7 @@
 import { withSentry } from '@sentry/nextjs'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
-import dbConnect from '../../../../lib/connectMongo'
+import dbConnect from '../../../../lib/mongoClient'
 import { IProject, Project } from '../../../../models/project.model'
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
